moved shared entity fields and functions to trait
