+++ config/packages/doctrine.yaml
@@ -12,7 +12,9 @@ doctrine:
             default:
                 url: '%env(resolve:DATABASE_URL)%'
                 driver: 'pdo_mysql'
-                server_version: '5.7'
+                # this setting prevents automatic database detection and finds a lot of false-negatives on doctrine:migrations:diff
+                # for null columns with MariaDB. Each migration tries to convert EVERY nullable column.
+                # server_version: '5.7'
                 charset: utf8mb4
                 default_table_options:
                     charset: utf8mb4
