+++ src/Entity/Project.php
@@ -83,21 +83,8 @@ class Project
      */
     private $activities;
 
-    /**
-     * @var float
-     *
-     * @ORM\Column(name="fixed_rate", type="float", precision=10, scale=2, nullable=true)
-     * @Assert\GreaterThanOrEqual(0)
-     */
-    private $fixedRate = null;
-
-    /**
-     * @var float
-     *
-     * @ORM\Column(name="hourly_rate", type="float", precision=10, scale=2, nullable=true)
-     * @Assert\GreaterThanOrEqual(0)
-     */
-    private $hourlyRate = null;
+    // keep the trait include exactly here, for placing the column at the correct position
+    use RatesTrait;
 
     /**
      * @var Timesheet[]
@@ -262,44 +249,6 @@ class Project
     public function setOrderNumber($orderNumber)
     {
         $this->orderNumber = $orderNumber;
-
-        return $this;
-    }
-
-    /**
-     * @return float
-     */
-    public function getFixedRate(): ?float
-    {
-        return $this->fixedRate;
-    }
-
-    /**
-     * @param float $fixedRate
-     * @return Project
-     */
-    public function setFixedRate(?float $fixedRate)
-    {
-        $this->fixedRate = $fixedRate;
-
-        return $this;
-    }
-
-    /**
-     * @return float
-     */
-    public function getHourlyRate(): ?float
-    {
-        return $this->hourlyRate;
-    }
-
-    /**
-     * @param float $hourlyRate
-     * @return Project
-     */
-    public function setHourlyRate(?float $hourlyRate)
-    {
-        $this->hourlyRate = $hourlyRate;
 
         return $this;
     }
