+++ src/Entity/RatesTrait.php
@@ -1,0 +1,70 @@
+<?php
+
+/*
+ * This file is part of the Kimai time-tracking app.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
+
+trait RatesTrait
+{
+    /**
+     * @var float
+     *
+     * @ORM\Column(name="fixed_rate", type="float", precision=10, scale=2, nullable=true)
+     * @Assert\GreaterThanOrEqual(0)
+     */
+    private $fixedRate = null;
+
+    /**
+     * @var float
+     *
+     * @ORM\Column(name="hourly_rate", type="float", precision=10, scale=2, nullable=true)
+     * @Assert\GreaterThanOrEqual(0)
+     */
+    private $hourlyRate = null;
+
+    /**
+     * @return float
+     */
+    public function getFixedRate(): ?float
+    {
+        return $this->fixedRate;
+    }
+
+    /**
+     * @param float $fixedRate
+     * @return self
+     */
+    public function setFixedRate(?float $fixedRate)
+    {
+        $this->fixedRate = $fixedRate;
+
+        return $this;
+    }
+
+    /**
+     * @return float
+     */
+    public function getHourlyRate(): ?float
+    {
+        return $this->hourlyRate;
+    }
+
+    /**
+     * @param float $hourlyRate
+     * @return self
+     */
+    public function setHourlyRate(?float $hourlyRate)
+    {
+        $this->hourlyRate = $hourlyRate;
+
+        return $this;
+    }
+}
